requirements.txt
yfinance
lxml
python-dotenv
newsapi-python